+++ bootique-jersey-client/src/main/java/io/bootique/jersey/client/auth/BasicAuthenticatorFactory.java
@@ -7,7 +7,6 @@ import io.bootique.annotation.BQConfigProperty;
 
 import javax.ws.rs.client.ClientRequestContext;
 import javax.ws.rs.client.ClientRequestFilter;
-import javax.ws.rs.core.Configuration;
 import javax.ws.rs.core.MultivaluedMap;
 import javax.xml.bind.DatatypeConverter;
 import java.io.IOException;
@@ -43,7 +42,7 @@ public class BasicAuthenticatorFactory implements AuthenticatorFactory {
 	}
 	
 	@Override
-	public ClientRequestFilter createAuthFilter(Configuration clientConfig, Injector injector) {
+	public ClientRequestFilter createAuthFilter(Injector injector) {
 		Objects.requireNonNull(username);
 		Objects.requireNonNull(password);